move model files
